Spelling error in JS node
So I had created a notification in the flow under "Recv access event over MQTT".  It followed directly after "Check if user is known."  The notification was being sent twice with two different messages.  Then I noticed that line 12 of the Javascript node was supposed to define msg.payload as "" but was spelled msg.paylod, which means the `return msg;` line was sending the "adding event" information in addition to the access information.  I changed it to payload and tested it and it works.